fix: cr issues
