+++ packages/site/examples/algorithm/changePosition/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "gaddi.js",
+      "filename": "pattern-matching.js",
       "title": {
         "zh": "图模式匹配",
         "en": "Graph pattern matching"
@@ -13,7 +13,7 @@
       "screenshot": "https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*uVOySpSIZXkAAAAAAAAAAAAADmJ7AQ/original"
     },
     {
-      "filename": "shortestPath.js",
+      "filename": "shortest-path.js",
       "title": {
         "zh": "最短路径",
         "en": "Shortest path"
